First Python Project
This is the first python project I started working on. I would really like to get inputs from others on how can I improve it or anything wrong with the code itself. I am trying to make a sort of accounts and a login program. I might also be adding different graphs to represent different things.